Re-factor of Sharing Schemas for ODM V2
Rewrote the specifications for the sharing process and sharing schemas to be inline with version 2 of the ODM.

Notable changes were the references to tables, headers, and categories using current terms, and inclusion of specific references to the "license" fields. Also the pivot from being an exclusion-based schema previously, and now being an inclusion-based schema.

Also corrected some internal inconsistencies and errors in the the text as I reviewed it.