Updates to sharing csv structure - Filter & Select
Rewrote the structure of the tables in line with the "filter" and "select" functions in Dplyr, trying to use the same structure.

Also updated accompanying text and some details about functionality in line with the changes to the table.

Also added information on a "sharing_metadata" table, based on Doug's comment above.