Update documentation/README with stable changes
After a long discussion on the best approach for the structure of the sharing schema csv, how best to operate the library, etc. It seems we've landed on a stable idea for the structure and how it will work - by generating SQL queries.

With this we should be able to close this PR and move to the next issue.