cli: use logging instead of stdout
to avoid spamming the terminal when running the tool.

The python logging doesn't write to stdout like print does, while still
allowing us to have a log of what happened (to a file, etc.). This is
good because we don't want to pollute stdout when making a CLI program.
Our errors are written to stderr instead, so that's fine.